refactor: break the mess of sub routing out into a hook
